video-to-docs/requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
